Removed two compilation warnings

+++ Tests/NumLib/NaturalNodeCoordinates.cpp
@@ -58,7 +58,7 @@ bool test(MeshLib::Element const& element)
             shape_matrices, GlobalDim, false /* axial symmetry */);
 
         auto const& N = shape_matrices.N;
-        for (int p = 0; p < ShapeFunction::NPOINTS; ++p)
+        for (int p = 0; p < static_cast<int>(ShapeFunction::NPOINTS); ++p)
         {
             if (p == n)
             {
